feature(da): Rust-based protobuf-like protocol (#945)

+++ nomos-da/kzgrs-backend/benches/verifier.rs
@@ -35,7 +35,7 @@ fn verify<const SIZE: usize>(bencher: Bencher, column_size: usize) {
             let sk = SecretKey::key_gen(&buff, &[]).unwrap();
             let verifier = DaVerifier::new(
                 sk.clone(),
-                (0..column_size as u32).collect(),
+                (0..column_size as u16).collect(),
                 GLOBAL_PARAMETERS.clone(),
             );
             let da_blob = DaBlob {
